// src/model/stat-name.type.ts
export type StatName =
	| 'totalDamageDealtToMinions'
	| 'totalDamageTakenByMinions'
	| 'totalDamageDealtToHeroes'
	| 'maxDamageDealtToHero'
	| 'highestWinStreak'
	| 'triplesCreated'
	| 'maxBoardStats'
	| 'coinsWasted'
	| 'rerolls'
	| 'freezes'
	| 'heroPowerUsed'
	| 'minionsBought'
	| 'minionsSold'
	| 'enemyMinionsKilled'
	| 'enemyHeroesKilled'
	| 'percentageOfBattlesGoingFirst'
	| 'battleLuck'
	| 'negativeBattleLuck';